test(accounts): name the test files after the modules they mirror
CI's real failure, and it was mine — not the inherited leak I first blamed and
not the environment-dependent assertion I fixed next:

    [PS-204 §2 orphan-test-file] tests/.../cli_pkg/test__account_cli_surface.py:
      no matching src file (orphan test); mirror dir src/.../cli_pkg contains
      ... none match expected `_account_cli_surface.py`

I named one file after a CONCEPT ("the cli surface") rather than after a module.
The mirror rule maps test__X.py -> src/.../_X.py, and there is no
_account_cli_surface.py. The change touches TWO modules, so it is two files:

    test__account_keepalive_cli_surface.py -> _account_keepalive.py   (12 tests)
    test_account_group_categories.py       -> account_group.py         (3 tests)

The convention was visible in the neighbours the whole time — eight existing
test_account_group_*.py files and test__account_keepalive_optional_peer.py. I
had read one of them earlier today and still invented my own name.

SAME RULE, SECOND TIME TODAY. PS-204 §2 also caught
tests/.../\_helpers/test_thread_leak.py on the thread-leak branch an hour ago.
Twice in one session I put a test where its name did not mirror a module. The
rule is not obscure; it is that I name files for what they are ABOUT instead of
what they COVER, and the audit is the only thing that notices.

VERIFIED against the check that actually failed, not a proxy:
    tests/develop/test_audit.py::test_audit_all_clean                    PASS
    tests/develop/test_tests_mirror_src.py::test_every_test_dir_mirrors  PASS
    2 passed in 71.59s
    the two split files: 15 collected / 15 passed / 0 failed

No test was dropped in the split — the script refused to write unless the two
halves summed to 15.
